Update patch set 3
Patch Set 3:

> I tried this before and it caused extreme ueventd selinux spam, so
 > the way we have it now was to workaround that. does the issue still
 > occur?

The selinux uevent spam is because cpu topology is disabled when the core is off and uevent is trying to set permission for a folder that doesn't exists. When the core is back on the /topology dir reappears and thus uevent can set permissions back again. This is not related to how qcom.power structure is.

Patch-set: 3
Reviewer: Keita Espinoza <15257@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0