Update patch set 1
Patch Set 1:

I propose to mark that topic as DNM as it is more like a work around and does not address the root cause of echos. The hardware of the z3 and z3c is suitable for active noise cancellation while on speaker mode during calls, and it is enabled on stock ROMs as well. So we should be able to keep it enabled as well.

Patch-set: 1
Reviewer: Gerrit User 24529 <24529@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>